fix setof to use catch instead of trying to do its own thing; fix unnecessary white lines when outputting solutions.
git-svn-id: https://yap.svn.sf.net/svnroot/yap/trunk@788 b08c6af1-5177-4d33-ba66-4b1c6b8b522a

+++ C/arith2.c
@@ -313,7 +313,7 @@ p_fdiv(Term t1, Term t2 E_ARGS)
       {
 	Int i1 = IntegerOfTerm(t1);
 	Float f2 = mpz_get_d(Yap_BigIntOfTerm(t2));
-	RFLOAT(i1/f2);
+	RFLOAT(((Float)i1/f2));
       }
 #endif
     default:
